+++ crates/dslab-dag/src/schedulers/heft.rs
@@ -108,7 +108,7 @@ impl HeftScheduler {
                     task_id,
                 ));
             }
-            memory_usage[best_resource].add(best_start, best_finish, dag.get_task(task_id).memory as i64);
+            memory_usage[best_resource].add(best_start, best_finish, dag.get_task(task_id).memory);
             task_finish_times[task_id] = best_finish;
             result.push((
                 best_start,
